refactor(mcp): consolidate MCP architecture into module/mcp directory
This commit reorganizes the MCP architecture files to consolidate
them into a dedicated module/mcp directory, improving code
organization and separation of concerns.

## Changes

### Core MCP Architecture
- Moved internal/server/mcp/ → internal/server/module/mcp/
- Unified FormatAdapter implementations
- Generic processing engines (loop + streaming)
- Forwarding layer abstraction

### File Organization
- **Module Components**: Core MCP types and interfaces now in module/mcp/
- **Server Integration**: Helper files remain in internal/server/
- **Test Files**: E2E tests kept at server level for integration testing

### Package Structure
- `module/mcp`: package mcp (core MCP abstractions)
- `internal/server`: package server (integration layer)
- Clean separation: core MCP logic vs server integration

### Updated Import Paths
- Updated import paths in helper and test files
- Maintained backward compatibility
- No breaking changes to public APIs

## Test Coverage
- ✅ All MCP path matrix tests passing (7/7)
- ✅ E2E test coverage maintained
- ✅ Integration tests validated

This reorganization improves code discoverability and makes
the MCP module boundaries clearer.
