+++ internal/server/module/mcp/anthropic_beta_adapter.go
