+++ internal/server/module/mcp/anthropic_v1_adapter.go
