+++ internal/server/module/mcp/format_adapter.go
