+++ internal/server/module/mcp/generic_loop_processor.go
