+++ internal/server/module/mcp/generic_stream_interceptor.go
