+++ internal/server/module/mcp/generic_stream_interceptor_test.go
