+++ internal/server/module/mcp/mcp_e2e_test.go
