+++ internal/server/module/mcp/mock_adapter_test.go
