+++ internal/server/module/mcp/openai_chat_adapter.go
